push 14/08.2025

+++ quarkus-service-classe-personnel/src/main/java/org/acme/aws/reactive_quarkus/simple/entity/SchoolYear.java
@@ -10,14 +10,14 @@ public class SchoolYear { // extends PanacheEntityBase
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Long id;
+    public Integer id;
     
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
